+++ src/BigCommerce/Post_Types/Product/Query.php
@@ -128,6 +128,37 @@ class Query {
 					};
 					add_filter( 'posts_orderby', $orderby_filter, 10, 2 );
 					break;
+					
+				case Product_Archive::SORT_INVENTORY_COUNT:
+					$meta_query = $query->get( 'meta_query' ) ?: [];
+
+					$meta_query['bigcommerce_inventory_level'] = [
+						'key'     => Product::INVENTORY_META_KEY,
+						'compare' => 'EXISTS',
+						'type'    => 'NUMERIC'
+					];
+
+					$meta_query['bigcommerce_id'] = [
+						'key'  => Product::BIGCOMMERCE_ID,
+						'type' => 'NUMERIC'
+					];
+
+					$query->set( 'meta_query', $meta_query );
+					$query->set( 'orderby', [ 'bigcommerce_inventory_level' => 'DESC', 'bigcommerce_id' => 'DESC' ] );
+					break;
+					
+				case Product_Archive::SORT_SKU:
+					$meta_query = $query->get( 'meta_query' ) ?: [];
+
+					$meta_query['bigcommerce_sku_normalized'] = [
+						'key'     => Product::SKU_NORMALIZED,
+						'compare' => 'EXISTS',
+					];
+
+					$query->set( 'meta_query', $meta_query );
+					$query->set( 'orderby', [ 'bigcommerce_sku_normalized' => 'ASC'] );
+
+					break;
 				default:
 					do_action( 'bigcommerce/query/sort', $query );
 					break;